+++ src/kmad.cpp
@@ -16,9 +16,9 @@ namespace po = boost::program_options;
 
 int main(int argc, char *argv[]) {
     int codon_length = 0;
-    int ptm_modifier = 0;
-    int domain_modifier = 0;
-    int motif_modifier = 0;
+    double ptm_modifier = 0;
+    double domain_modifier = 0;
+    double motif_modifier = 0;
     double gap_ext_pen = 0;
     double gap_open_pen;
     double end_pen = 0;
@@ -53,13 +53,13 @@ int main(int argc, char *argv[]) {
        po::value<int>(&codon_length)->implicit_value(7)
                                     ->default_value(1),"codon length")
       ("phosph,p",
-       po::value<int>(&ptm_modifier)->default_value(10),
+       po::value<double>(&ptm_modifier)->default_value(10),
        "score for aligning phosphorylated residues")
       ("domain,d",
-       po::value<int>(&domain_modifier)->default_value(3),
+       po::value<double>(&domain_modifier)->default_value(3),
        "score for aligning domains")
       ("motif,m",
-       po::value<int>(&motif_modifier)->default_value(3),
+       po::value<double>(&motif_modifier)->default_value(3),
                                  "probability multiplier for motifs")
       ("out-encoded",
        po::value<bool>(&out_encoded)->implicit_value(true)
